# Conflicts:
#	configure.ac

+++ src/spork.cpp
@@ -95,9 +95,9 @@ bool IsSporkActive(int nSporkID)
 }
 
 // grab the value of the spork on the network, or the default
-int GetSporkValue(int nSporkID)
+long GetSporkValue(int nSporkID)
 {
-    int r = 0;
+    long r = 0;
 
     if(mapSporksActive.count(nSporkID)){
         r = mapSporksActive[nSporkID].nValue;
